capstone_slack/requirements.txt
slack-sdk
Flask
slackeventsapi
requests
python-dotenv
